+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Unified the `Verified` and `Signed` token header types, since they represent the same state – a known, validated signature and related parameters are present.
+
 ## [1.0.1](https://github.com/alexrudy/jaws/releases/tag/1.0.1)
 
 Every patch will bump dependencies as appropriate.